Fixed case of the filename
