+++ storybuilder/Frontend/public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" data-mantine-color-scheme="dark">
   <head>
     <meta charset="utf-8" />
     <link rel="icon" href="./favicon.ico" />
