--- storybuilder/Frontend/src/App.css
@@ -1,121 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-html {
-    font-size: 62.5%;
-    overflow: hidden;
-}
-
-body {
-    background: rgb(6, 6, 65);
-    color: white;
-    font-family: sans-serif;
-    overflow: hidden;
-}
-
-.App {
-    min-height: 100vh;
-    display: flex;
-}
-
-.sidebar {
-    flex: 2;
-    border-right: 1px solid rgb(100, 100, 100);
-    padding: 2.5rem;
-    display: flex;
-    align-items: flex-start;
-}
-
-.logo {
-    display: flex;
-    align-items: center;
-}
-
-.logo-name {
-    font-size: 2.5rem;
-    margin-left: 1rem;
-}
-
-.main {
-    flex: 9;
-    min-height: calc(100vh - 14rem);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 6rem 10rem;
-    margin-bottom: 3%;
-}
-
-.chat-icons {
-    width: 4rem;
-    height: 4%;
-    object-fit: cover;
-    border-radius: 0.5rem;
-    margin-right: 2rem;
-}
-
-.chats {
-    overflow: hidden;
-    overflow-y: auto;
-    scroll-behavior: smooth;
-    width: 100%;
-    max-width: 120rem;
-    height: calc(100vh - 17rem);
-    scrollbar-width: thin;
-    scrollbar-color: #888 transparent;
-}
-
-.chat {
-    margin: 1rem;
-    padding: 2rem 3rem;
-    font-size: 1.2rem;
-    display: flex;
-    align-items: flex-start;
-    text-align: justify;
-}
-
-.ai {
-    background: rgba(28, 30, 58, 1);
-    width: fit-content;
-    border-radius: 0.5rem;
-}
-
-.chat-footer {
-    margin-top: auto;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.input-box {
-    padding: 0.5rem;
-    background: rgba(28, 30, 58, 1);
-    display: flex;
-    align-items: center;
-    border-radius: 0.5rem;
-    width: 120rem;
-}
-
-.input-box>input {
-    width: calc(100% - 3rem);
-    outline: none;
-    padding: 1.25rem;
-    color: white;
-}
-
-.input-box>input, .send-button {
-    background: transparent;
-    border: none;
-}
-
-.send-icon {
-    width: 2.5rem;
-    height: 2.5rem;
-    object-fit: cover;
-    border-radius: 1rem;
-}
