+++ storybuilder/Frontend/src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import { MantineProvider } from '@mantine/core';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <App />
+    <MantineProvider theme={{colorScheme: 'dark'}} defaultColorScheme="dark" withGlobalStyles withNormalizeCSS>
+        <App />
+    </MantineProvider>
 );
